+++ pkg/cortex/cortex.go
@@ -22,6 +22,7 @@ import (
 	"gopkg.in/yaml.v2"
 
 	"github.com/cortexproject/cortex/pkg/util/grpcclient"
+	"github.com/cortexproject/cortex/pkg/util/resource"
 
 	"github.com/cortexproject/cortex/pkg/alertmanager"
 	"github.com/cortexproject/cortex/pkg/alertmanager/alertstore"
@@ -88,10 +89,11 @@ var (
 
 // Config is the root config for Cortex.
 type Config struct {
-	Target      flagext.StringSliceCSV `yaml:"target"`
-	AuthEnabled bool                   `yaml:"auth_enabled"`
-	PrintConfig bool                   `yaml:"-"`
-	HTTPPrefix  string                 `yaml:"http_prefix"`
+	Target             flagext.StringSliceCSV `yaml:"target"`
+	AuthEnabled        bool                   `yaml:"auth_enabled"`
+	PrintConfig        bool                   `yaml:"-"`
+	HTTPPrefix         string                 `yaml:"http_prefix"`
+	MonitoredResources flagext.StringSliceCSV `yaml:"monitored_resources"`
 
 	ExternalQueryable prom_storage.Queryable `yaml:"-"`
 	ExternalPusher    ruler.Pusher           `yaml:"-"`
@@ -142,6 +144,11 @@ func (c *Config) RegisterFlags(f *flag.FlagSet) {
 	f.BoolVar(&c.AuthEnabled, "auth.enabled", true, "Set to false to disable auth.")
 	f.BoolVar(&c.PrintConfig, "print.config", false, "Print the config and exit.")
 	f.StringVar(&c.HTTPPrefix, "http.prefix", "/api/prom", "HTTP path prefix for Cortex API.")
+
+	c.MonitoredResources = []string{}
+	f.Var(&c.MonitoredResources, "monitored.resources", "Comma-separated list of resources to monitor. "+
+		"Supported values are cpu and heap, which tracks metrics from github.com/prometheus/procfs and runtime/metrics "+
+		"that are close estimates. Empty string to disable.")
 
 	c.API.RegisterFlags(f)
 	c.registerServerFlagsWithChangedDefaultValues(f)
@@ -216,7 +223,7 @@ func (c *Config) Validate(log log.Logger) error {
 	if err := c.QueryRange.Validate(c.Querier); err != nil {
 		return errors.Wrap(err, "invalid query_range config")
 	}
-	if err := c.StoreGateway.Validate(c.LimitsConfig); err != nil {
+	if err := c.StoreGateway.Validate(c.LimitsConfig, c.MonitoredResources); err != nil {
 		return errors.Wrap(err, "invalid store-gateway config")
 	}
 	if err := c.Compactor.Validate(c.LimitsConfig); err != nil {
@@ -229,12 +236,22 @@ func (c *Config) Validate(log log.Logger) error {
 		return errors.Wrap(err, "invalid alertmanager config")
 	}
 
-	if err := c.Ingester.Validate(); err != nil {
+	if err := c.Ingester.Validate(c.MonitoredResources); err != nil {
 		return errors.Wrap(err, "invalid ingester config")
 	}
 
 	if err := c.Tracing.Validate(); err != nil {
 		return errors.Wrap(err, "invalid tracing config")
+	}
+
+	for _, r := range c.MonitoredResources {
+		switch resource.Type(r) {
+		case resource.CPU, resource.Heap:
+		default:
+			if len(r) > 0 {
+				return fmt.Errorf("unsupported resource type to monitor: %s", r)
+			}
+		}
 	}
 
 	return nil
@@ -315,6 +332,7 @@ type Cortex struct {
 	MetadataQuerier          querier.MetadataQuerier
 	QuerierEngine            promql.QueryEngine
 	QueryFrontendTripperware tripperware.Tripperware
+	ResourceMonitor          *resource.Monitor
 
 	Ruler        *ruler.Ruler
 	RulerStorage rulestore.RuleStore